updates by Skala
merged:

  gentoo-x86: refs/heads/gentoo.org
  party-overlay: 8f06b1ad5eff6182f357dc8f5c710490f39cb218
  elementary: aaea54a97c83209ccee1d87299df5d8474772ee7
  funtoo-overlay: d9b809d83bc20b285c3aff7c8d06332fcdc4a73b
  sabayon-for-gentoo: d249757a914e22074b9e1cfeb89c61eae3bba4ac

+++ sys-kernel/debian-sources/debian-sources-3.16.7.ebuild
@@ -5,10 +5,10 @@ EAPI=5
 inherit check-reqs eutils mount-boot
 
 SLOT=$PVR
-CKV=3.16.7-ckt4
+CKV=3.16.7-ckt7
 KV_FULL=${PN}-${PVR}
-EXTRAVERSION=-ckt4-1
-KERNEL_ARCHIVE="linux_${PV}-ckt4.orig.tar.xz"
+EXTRAVERSION=-ckt7-1
+KERNEL_ARCHIVE="linux_${PV}-ckt7.orig.tar.xz"
 PATCH_ARCHIVE="linux_${PV}${EXTRAVERSION}.debian.tar.xz"
 RESTRICT="binchecks strip mirror"
 # based on : http://packages.ubuntu.com/maverick/linux-image-2.6.35-22-server
@@ -146,13 +146,13 @@ src_install() {
 }
 
 pkg_postinst() {
-	if [[ -h "${ROOT}"usr/src/linux ]]; 
-	then 
+	if use binary && [[ -h "${ROOT}"usr/src/linux ]]; then 
 		rm "${ROOT}"usr/src/linux
 	fi
 
-	if [[ ! -e "${ROOT}"usr/src/linux ]];
-	then
+	if use binary && [[ ! -e "${ROOT}"usr/src/linux ]]; then
+		ewarn "With binary use flag enabled /usr/src/linux"
+		ewarn "symlink automatically set to debian kernel"
 		ln -sf linux-${P} "${ROOT}"usr/src/linux
 	fi
